feat: 카테고리 관련 엔티티 정의
- TrashCategory
  - 일반쓰래기, 불가연성 등 큰 범주
- TrashSubCategory
  - 의자, 병, 휴대폰 등 물건의 종류에 대한 범주
- 위 엔티티는 소스코드에서 Create, Update, Delete 할 수 없도록 함
  - JPA를 위한 기본 생성자만 생성
  - JpaRepository가 아닌 Repository를 상속하게 하여 필요시 직접 작성하는 구조
- Trash의 카테고리 관련 멤버 삭제
