+++ src/main/java/store/sonyk9919/api/domain/disposal/entity/TrashDisposal.java
@@ -1,0 +1,20 @@
+package store.sonyk9919.api.domain.disposal.entity;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class TrashDisposal {
+
+    @Id
+    @Column(name = "disposal_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true, columnDefinition = "TEXT")
+    private String guide;
+}
