+++ src/main/java/store/sonyk9919/api/domain/taxonomy/entitiy/TrashCategory.java
@@ -1,0 +1,20 @@
+package store.sonyk9919.api.domain.taxonomy.entitiy;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class TrashCategory {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "category_id")
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String name;
+}
