+++ src/main/java/store/sonyk9919/api/domain/taxonomy/entitiy/TrashSubCategory.java
@@ -1,0 +1,23 @@
+package store.sonyk9919.api.domain.taxonomy.entitiy;
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class TrashSubCategory {
+
+    @Id
+    @Column(name = "subcategory_id")
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false, unique = true)
+    private String name;
+
+    @Column(nullable = false, unique = true)
+    private String alias;
+}
