+++ src/main/java/store/sonyk9919/api/domain/taxonomy/repository/TrashSubCategoryRepository.java
@@ -1,0 +1,7 @@
+package store.sonyk9919.api.domain.taxonomy.repository;
+
+import org.springframework.data.repository.Repository;
+import store.sonyk9919.api.domain.taxonomy.entitiy.TrashSubCategory;
+
+public interface TrashSubCategoryRepository extends Repository<TrashSubCategory, Long> {
+}
